requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
httpx==0.28.1
python-dotenv==1.0.1

# Anthropic Claude
anthropic==0.44.0

# MCP SDK
mcp==1.9.0

# FHIR
fhir.resources==8.0.0
fhirclient==4.2.0

# Drug Data
requests==2.32.3

# RAG / Guidelines
chromadb==0.6.3
sentence-transformers==3.4.1

# A2A Protocol
sse-starlette==2.2.1
uuid6==2024.7.10

# Utilities
loguru==0.7.3
tenacity==9.0.0
jinja2==3.1.5
pyyaml==6.0.2

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-httpx==0.35.0

# Dev
black==24.10.0
ruff==0.9.3
mypy==1.14.1
